feat: add FileDropUploader
