+++ packages/apps/spaces/ui/form/FileUploader/index.ts
@@ -1,1 +1,2 @@
-export * from './FileUploader';
+export * from './FileUploadTrigger';
+export * from './FileDropUploader';
